==> Regnskapsflyten/Balanseregnskapet/2. Kommune- og fylkeskommunekassa Balanse.ipynb <==
# KOSTRA skjema B og D


## Henter nødvendige pakker
**For løpet Kostra0B_0D trengs følgende pakker:**  
   - Pandas
   - Numpy
   - Selvlagde funksjoner lagret i fil funksjoner.py

import pandas as pd
import numpy as np
exec(open('../funksjoner.py').read())

## Angi årgang som produseres

#Denne bør legges i en funksjon. Så man kun trenger å skrive årgang()
aar=input('Angi årgang for filene:' )
aar_1t = str(int(aar) - 1)
aar_t1 = str(int(aar) + 1)
print("Du lager nå filer for " + aar)

## Henter nødvendige datafiler
**For løpet Kostra0B_0D trengs følgende datafiler:**
- Import fra Driller
    - Skjema KOSTRA0D
    - Skjema KOSTRA0D
- Lister fra KLASS
    - Kodeliste for Kostra balansekapitler
    - Kommunelisten for gjeldene årgang
    - Fylkeskommunelisten for gjeldene årgang

kostra0b = pd.read_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/import_driller/" + aar + "/kostra0b.parquet")
kostra0d = pd.read_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/import_driller/" + aar + "/kostra0d.parquet")
kap = pd.read_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/klass/" + aar + "/kapittel.parquet").astype(str)
komm = pd.read_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/klass/" + aar + "/komm.parquet").astype(str)
fykomm = pd.read_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/klass/" + aar + "/fykomm.parquet").astype(str)

## Bearbeider regnskapsfilene

### Slår sammen Kostra0B og Kostra0D

#Setter sammen skjema a og c
bd = pd.concat([kostra0b, kostra0d])

### Ordner data
- Definerer rett datatype
- Endrer kolonnenavn
    - Fra store til små bokstaver
    - Endrer navn
- Fjerner manglende observasjoner

bd = bd.astype({'FUNKSJON_KAPITTEL':str,'ART_SEKTOR':str, 'KONTOKLASSE':str,'PERIODE':str,'REGION':str, 'BYDEL':str})

#Endrer headingen fra store til små bokstaver
bd.columns = [x.lower() for x in bd.columns]

#Endrer navn på kapittel
bd.rename(columns={'funksjon_kapittel':'kapittel'}, inplace = True)

bd = bd.groupby(['kapittel', 'kontoklasse', 'periode', 'region', 'bydel'])[['belop']].sum().reset_index()

#Filtrerer bort alle rader med manglede verdier. Fungerer det? Må testes på et annet datasett som faktisk har manglende verdier... 
bd_null = bd.dropna()

#### Kontrollerer fjernede rader

kontroll_merge(bd,bd_null)

### Snur fortegn for inntektsarter
- Endrer alle verdier som har art høyere eller lik 600 til negative tall

bd_snudd = bd_null.copy()

bd_snudd['belop']=np.where((bd_snudd.kapittel >= '31') & (bd_snudd.kapittel <= '5990'),  bd_snudd.belop*-1, bd_snudd.belop)

## Lager bydelsregioner
- Slår sammen regions- og bydelsnummer til samme kolonne

#Henter ut alle Oslo tall som har bydel. Slår sammen region og bydelkode.
bydeler = bd_snudd[(bd_snudd['region'] == '0301') & (bd_snudd['bydel'] != '00')].copy()
bydeler["region"] = bydeler["region"] + bydeler["bydel"]

#Fjerner bydelene fra originalfilen, slår sammen den nye bydelsfilen og arbdeisfilen. Fjerner kolonnen for bydeler. 
bd_uten_bydel = bd_snudd[(bd_snudd['bydel'] == '00')].copy()
bd_bydeler = pd.concat([bd_uten_bydel, bydeler]).drop(columns=['bydel'])

## Lager Oslo Fylkeskommune
- Tallene fra Oslo skal være med både for kommune og fylkeskommune. Vi må derfor kopiere alle Oslo tall og endre region på de kopierte radene fra 0301 til 0300. 

#Henter ut Oslo og endrer regionskode til 0300. Legger Oslo fylkeskommune på arbeidsfilen. 
oslo=bd_bydeler[(bd_bydeler['region']=='0301')].copy()
oslo['region'] = oslo['region'].str.replace('0301', '0300')

- Slår sammen Oslo Fylkeskommune med regnskapsfilen

bd_forts = pd.concat([bd_bydeler, oslo])

## Filtrerer gyldige kapitler
**Kapittel listen**
- Hentes direkte fra KLASS via API i egen notebook (API Klass). 
- Filen laget i API-notebooken er lastet inn helt i starten av denne boken.

**Kontroll**
- Funksjonen kontroll_merge er laget i egen notebook (funksjoner.py)
- Kontrollen lister ut alle rader som er fjernet i mergen over. Dersom ingen rader listes ut ble ingenting fjernet i operasjonen over.

### Kapittelsfilter
- Merger alle gylige kapitler med regnskapsfilen. Dette fjerner alle reder med ugyldige kapitler

kap=kap.drop(columns={'periode'})

gyldige_kap = list(kap.columns.values)

k1 = bd_forts.set_index(gyldige_kap).index
k2 = kap.set_index(gyldige_kap).index
kapittelsfilter = bd_forts[k1.isin(k2)]

#### Kontrollerer fjernede rader
- Må kunne se nærmere på dataen

kontroll_merge(bd_forts, kapittelsfilter)

# Foreløpig Fil - kontrollert

kostra0b_0d_kontrollert = kapittelsfilter

### Lagrer filen som skal benyttes i neste års imputering

kostra0b_0d_kontrollert.to_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/arbeidsfiler/" + aar + "/kostra0b_0d_til_imputering_" + aar + ".parquet")

## Finner manglende kommuner
- For å finne hvem som mangler må vi koble opp mot regionslisten. Dette gjøres via API mot Klass (notebook API Klass).

- Lagt opp mot kommuner og fylkeskommuner, ikke bydeler


#Lister ut alle kommuner og fylkeskommuner som er med i regnskapsfilen. 
#Lagrer filen til senere bruk.
levert = kostra0b_0d_kontrollert[['region']].drop_duplicates()
levert.to_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/arbeidsfiler/" + aar + "/kostra0b_0d_levert_" + aar + ".parquet")

### Alle regioner

#Slår sammen kommujne og fylkeskommunelisten tidligere hentet fra Klass. 
regioner = pd.concat([komm, fykomm])

### Manglende kommuner

#Lister ut manglende kommuner
manglende = kontroll_merge(regioner,levert)
manglende

manglende.to_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/arbeidsfiler/" + aar + "/kostra0b_0d_manglende_" + aar + ".parquet")

## Imputerer for manglende kommuner

### Henter og tilpasser fjorårets regnskap

#Henter inn fjorårets 0A_0C-fil
b_d_1t = pd.read_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/arbeidsfiler/" + aar_1t + "/kostra0b_0d_til_imputering_" + aar_1t + ".parquet")

- Endrer årgang

b_d_1t.loc[b_d_1t['periode'] == aar_1t, 'periode'] = aar

- Filtrerer fjorårets regnskap mot årets kapitteliste

k3 = b_d_1t.set_index(gyldige_kap).index
k2 = kap.set_index(gyldige_kap).index
estimat_kapittelsfilter = b_d_1t[k3.isin(k2)]

bd_1t = estimat_kapittelsfilter

- Filterer ut mangldene kommuner  

manglende_estimat=pd.merge(manglende, bd_1t, on=['region', 'periode']).drop(columns=['_merge'])

- Slår sammen årets regnskap med fjorårets for de manglende kommiunene

bd_basis = pd.concat([kostra0b_0d_kontrollert, manglende_estimat])

## Oppretter regnskapsomfang
- Regnskapsomfang B definerer at dette er kassetall

bd_basis['regnskapsomfang'] = 'B'

# Ferdig fil
- Lagrer fil som er klar som input til KostraRegnskaps-metode
- Her må vi ha mulighet for å se på dataen

bd_basis.to_parquet("/ssb/stamme03/komakro/pilot_python/arbeidsdata/arbeidsfiler/" + aar + "/kostra0b_0d_basis_" + aar + ".parquet")

#### Kontrollfilter

#sjekk=bd_basis[(bd_basis['region']=='4601') & (bd_basis['art']=='905') ].copy()
#sjekk

bd_basis
